+++ web/wp-content/plugins/groupsengine_plugin/groupsengine_plugin.php
@@ -2,7 +2,7 @@
 /* Plugin Name: Groups Engine 
 Plugin URI: http://groupsengine.com
 Description: Groups Engine is powerful group search that actually works. To get started, activate the plugin and open the new "Groups Engine" menu. Follow the instructions on the <a href="admin.php?page=groupsengine_plugin/groupsengine_plugin.php_userguide">User Guide page</a> to embed a groups page, change the color scheme and more.
-Version: 1.3.3
+Version: 1.3.4
 Author: Eric Murrell (Volacious) 
 Author URI: http://groupsengine.com */ 
 
@@ -11,14 +11,13 @@ Author URI: http://groupsengine.com */
 global $wp_version;
 /* ----- Plugin Updates ----- */
 require 'plugin-updates/plugin-update-checker.php';
-$ENMGEUpdateChecker = PucFactory::buildUpdateChecker(
+$ENMGEUpdateChecker = Puc_v4_Factory::buildUpdateChecker(
 	'http://pluginupdates.groupsengine.com/newupdater.json',
 	__FILE__,
 	'groupsengine_plugin'
 );
-
 /* ----- Install the Plugin ----- */
-define ( 'ENMGE_CURRENT_VERSION', '1.3.3' );
+define ( 'ENMGE_CURRENT_VERSION', '1.3.4' );
 
 $enmge_options = get_option( 'enm_groupsengine_options' );
 
@@ -682,7 +681,7 @@ function groupsengine_ajaxcontactfindgroup() { // Sort Files
 add_action( 'wp_ajax_groupsengine_ajaxcontactfindgroup', 'groupsengine_ajaxcontactfindgroup' );
 
 /* Refresh styles and options on plugin update */
-if ( get_option( 'enmge_db_version' ) && get_option( 'enmge_db_version' ) < "1.3.3" ) {
+if ( get_option( 'enmge_db_version' ) && get_option( 'enmge_db_version' ) < "1.3.4" ) {
 	include('includes/core/updates.php');
 }