+++ web/wp-content/plugins/typehub/typehub.php
@@ -9,7 +9,7 @@
  * Plugin Name:       Type Hub
  * Plugin URI:        http://brandexponents.com
  * Description:       Typehub lets you take complete control over the typography of your website.
- * Version:           2.0.5
+ * Version:           2.0.6
  * Author:            Brand Exponents
  * Author URI:        http://brandexponents.com
  * License:           GPL-2.0+
@@ -28,7 +28,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'TYPEHUB_VERSION', '2.0.5' );
+define( 'TYPEHUB_VERSION', '2.0.6' );
 
 if( !defined( 'TYPEHUB_PLUGIN_URL' ) ) {
 	define( 'TYPEHUB_PLUGIN_URL', plugins_url( '', __FILE__ ) );
@@ -60,7 +60,7 @@ register_deactivation_hook( __FILE__, 'deactivate_typehub' );
 
 require TYPEHUB_PLUGIN_DIR. 'plugin-update-checker/plugin-update-checker.php';
 $typehub_update_checker = new PluginUpdateChecker_3_1 (
-    'https://brandexponents.com/oshin-plugins/typehub.json',
+    'https://brandexponents.com/wp/wp-content/uploads/typehub.json',
     __FILE__,
     'typehub'
 );