+++ web/wp-content/themes/oshin/ReduxFramework/ReduxCore/inc/class.redux_filesystem.php
@@ -26,12 +26,14 @@
             public $parent = null;
 
             public function __construct() {
+              if (! empty($this->parent)) {
                 $this->parent->admin_notices[] = array(
-                    'type'    => 'error',
-                    'msg'     => '<strong>' . __( 'File Permission Issues', 'redux-framework' ) . '</strong><br/>' . sprintf( __( 'We were unable to modify required files. Please check your permissions, or modify your wp-config.php file to contain your FTP login credentials as <a href="%s" target="_blank">outlined here</a>.', 'redux-framework' ), 'https://codex.wordpress.org/Editing_wp-config.php#WordPress_Upgrade_Constants' ),
-                    'id'      => 'redux-wp-login',
-                    'dismiss' => false,
+                  'type'    => 'error',
+                  'msg'     => '<strong>' . __( 'File Permission Issues', 'pinnacle' ) . '</strong><br/>' . sprintf( __( 'We were unable to modify required files. Please check your permissions, or modify your wp-config.php file to contain your FTP login credentials as <a href="%s" target="_blank">outlined here</a>.', 'pinnacle' ), 'https://codex.wordpress.org/Editing_wp-config.php#WordPress_Upgrade_Constants' ),
+                  'id'      => 'redux-wp-login',
+                  'dismiss' => false,
                 );
+              }
             }
 
             /**
@@ -125,7 +127,7 @@
 
                 // Setup the filesystem with creds
                 require_once ABSPATH . '/wp-admin/includes/template.php';
-               
+
 		require_once ABSPATH . '/wp-includes/pluggable.php';
                 require_once ABSPATH . '/wp-admin/includes/file.php';
 
@@ -254,18 +256,18 @@
 			if (empty($res) || $res == false || $res == '' ) {
                             return;
 			}
-                        
+
                         if (is_array($res) && empty($res)) {
                             return;
                         }
-                        
+
                         if (!is_array($res)) {
                             if (count(glob("$file*")) == 0) {
                                 return;
                             }
                         }
                     }
-                    
+
                     $this->killswitch              = true;
                     $this->parent->admin_notices[] = array(
                         'type'    => 'error',