+++ web/wp-content/plugins/be-gdpr/changelog.md
@@ -1,3 +1,7 @@
+v1.1.4
+======
+* Fixed : Plugin update checker url
+
 v1.1.3
 ======
 * Fixed: bugs