+++ web/wp-content/plugins/typehub/changelog.md
@@ -1,9 +1,12 @@
 Changelog
 ========
+v2.0.6
+=====
+* Fixed: Plugin update checker url
 
 v2.0.5
 =====
-* Fixed : File name conatining spaces in custom fonts
+* Fixed: File name conatining spaces in custom fonts
 
 v2.0.4
 =====